Move client implementation to internal.pkg.client
Refactor: extract and relocate the HTTP client implementation from internal.core_python.client to a new internal.pkg.client package (including session, bootstrap and utils). Update all imports/usages across the codebase and tests to reference the new package, adjust typing modules accordingly, and add new pkg typing stubs. Add/rename login helpers (login_raw / login_interactive) and wire them through CLI and init code. Update top-level exports (sdk and package __init__ files) to expose helper/console/fs/safe and updated run/login APIs, and use helper.get_swanlab_version for __version__. Minor cleanup: fix safe package import, add __all__ entries, and move/rename multiple tests to match the new module layout.
