+++ swanlab/sdk/cmd/__init__.py
@@ -4,3 +4,27 @@
 @time: 2026/3/6 21:58
 @description: 封装API，提供给外部使用
 """
+
+from .init import init
+from .login import login, login_interactive
+from .logout import logout
+from .merge_settings import merge_settings
+from .run import async_log, define_scalar, finish, log, log_audio, log_image, log_text, log_video
+from .verify import verify
+
+__all__ = [
+    "init",
+    "verify",
+    "merge_settings",
+    "logout",
+    "login",
+    "login_interactive",
+    "log",
+    "log_text",
+    "log_image",
+    "log_audio",
+    "log_video",
+    "async_log",
+    "define_scalar",
+    "finish",
+]
