+++ swanlab/sdk/internal/core_python/api/project.py
@@ -9,8 +9,8 @@ from typing import Optional, cast
 
 from swanlab.exceptions import ApiError
 from swanlab.sdk.internal.core_python import client
-from swanlab.sdk.internal.core_python.client.utils import decode_response
 from swanlab.sdk.internal.pkg import helper
+from swanlab.sdk.internal.pkg.client.utils import decode_response
 from swanlab.sdk.typings.core_python.api.project import InitProjectType, ProjectType
