+++ swanlab/sdk/typings/pkg/__init__.py
@@ -1,0 +1,6 @@
+"""
+@author: cunyue
+@file: __init__.py
+@time: 2026/4/14 00:48
+@description: Pkg 模块类型提示
+"""
