+++ swanlab/sdk/typings/pkg/client/__init__.py
@@ -1,0 +1,6 @@
+"""
+@author: cunyue
+@file: __init__.py
+@time: 2026/4/14 00:48
+@description: SwanLab 客户端模块的类型提示定义
+"""
