+++ tests/unit/sdk/internal/pkg/client/test_client_session.py
@@ -9,7 +9,7 @@ import pytest
 import responses
 
 from swanlab.exceptions import ApiError
-from swanlab.sdk.internal.core_python.client.session import (
+from swanlab.sdk.internal.pkg.client.session import (
     TimeoutHTTPAdapter,
     create,
     request_retries_ctx,
